MFC r304543: Unbreak sctp_connectx().
MFC r304573:
Remove duplicate code, which is not protected by the appropriate locks.

MFC r304579:
Improve the locking when sending user messages.

First, keep a ref count on the stcb after looking it up, as
done in the other lookup cases.
Second, before looking again at sp, ensure that it is not
freed, because the assoc is about to be freed.


git-svn-id: https://svn.freebsd.org/base/stable/11@304792 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f