MFC r304953:
Define ipfilter's SOLARIS macro in a defined and portable way.

Reviewed by:	cy
Differential Revision: https://reviews.freebsd.org/D7671

MFC r304959 (by kib):

Complete r304953.

Sponsored by:	The FreeBSD Foundation

MFC r304964:

Follow-up to r304953, in which I broke the build: apparently the SOLARIS
macro is defined in lots of different places in ipfilter, so replace all
of the nonportable definitions with portable ones.

Pointy hat to:	dim


git-svn-id: https://svn.freebsd.org/base/stable/10@305138 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f