cache: manage negative entry list with a dedicated lock
Since negative entries are managed with a LRU list, a hit requires a
modificaton.

Currently the code tries to upgrade the global lock if needed and is
forced to retry the lookup if it fails.

Provide a dedicated lock for use when the cache is only shared-locked.

Reviewed by:	kib
MFC after:	1 week


git-svn-id: https://svn.freebsd.org/base/head@305379 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f