MFC r305265:
Revert r304192 to fix short month names and replace %b with %_m in date_fmt
for Chinese locales.

As mentioned in the commit message of r289041, nl_langinfo(ABMON_*) only
returned numbers when using a Chinese locale, this causes problems in
applications that put the short month name and the day of the month together.

Spotted by:	Ting-Wei Lan <lantw44 gmail com>


git-svn-id: https://svn.freebsd.org/base/stable/11@305452 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f